Rework the build system for a new tests approach
Perform a substantial cleanup of the build system with respect to
building and installing testcases.

By default, tests are no longer compiled as part of 'make'.  They will
be built when 'make check' is run.  The old behaviour can be obtained
with --enable-always-build-tests.

--disable-modular-tests is gone (because tests are no longer built by
default).  There is no longer any way to cause 'make check' to be a
no-op, but that's not very useful anyway.

A new glibtests.m4 file is introduced.  Along with a substantially
expanded Makefile.decl, this provides for consistent handling of
--enable-installed-tests and --enable-always-build-tests (mentioned
above).

Port our various test-installing Makefiles to the new framework.  This
is a net reduction of about ~350 lines of Makefile and configure.ac.

This patch substantially improves the situation in the toplevel tests/
directory.  Things are now somewhat under control there.  There were
some tests being built that weren't even being run and we run those now.
The long-running GObject performance tests in this directory have been
removed from 'make check' because they take too long.

There is one regression in this patch: the appinfo tests can no longer
be installed due to an inability to handle differently-generated data
files for installed vs. uninstalled tests.  I could probably work around
this, but I lack the will to do so now and I think we may be able to fix
this in GLib itself by introducing proper support for relative pathnames
to desktop files.