MSVC 2008 Projects: Split the Property Sheets
Split the property sheet into four sheets, to make maintainance of the
build files easier, and also to prepare for using autotools to fill in
the parts for the "installation" of headers.

Also put more of the items that are repeated in the projects into the
property sheets, also to improve ease of maintenance.

Updates to the Visual Studio 2010 projects will come later, as the script
to update them to Visual Studio 2012 must also be taken into account during'
the process.