gsubprocess: Cleanup GSubprocess correctly
Although it's nice that GSubprocess object lives until its
child goes away, this can't be the case during cleanup and or
unloading of gio, since the GLib worker thread will run non-existant
code.

So in the case that GIO hits cleanup, we finalize the GSubprocess
regardless of the state of its child.

https://bugzilla.gnome.org/show_bug.cgi?id=711799